Replace randcomb2 by randocomb3
Replace random combination algorithm (given a set of probabilities) from
the randcomb2 which was only easy to implement for RF=N-1, to a much more
general and simpler randcomb3 based on Systematic Sampling which makes fewer
guarantees but still guarantees exactly what we need.

Also added extensive comments in the code to explain what this means.