overhaul
Big overhaul, moving most of the code from prob.hh to prob.cc, removing
dead code, adding better comments, reorganizing stuff.

The API did not change, except for one thing: if me >= rf, this means this
node is not one of the replicas, and we don't need to do the redistribution
algorithm for prefering the local replica - but still need to calculate and
use the probabilities to send to the different nodes.